// src/app/core/navigation/navigation.types.ts
import { FuseNavigationItem } from '@fuse/components/navigation';

export interface Navigation {
    compact: FuseNavigationItem[];
    default: FuseNavigationItem[];
    futuristic: FuseNavigationItem[];
    horizontal: FuseNavigationItem[];
    accounting: FuseNavigationItem[];
    inventory: FuseNavigationItem[];
    sales: FuseNavigationItem[];
    purchases: FuseNavigationItem[];
    admin: FuseNavigationItem[];
    manufacturing: FuseNavigationItem[];
    hr: FuseNavigationItem[];
    settings: FuseNavigationItem[];
    logistics: FuseNavigationItem[];
    gate: FuseNavigationItem[];
    pos: FuseNavigationItem[];
}